flask/requirements.txt
alembic==1.4.3
astroid==2.4.2
Brotli==1.0.9
bson==0.5.10
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Compress==1.8.0
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.1.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.2.0
passlib==1.7.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
